list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -2,33 +2,31 @@
 
 /**
  * print_list - function that print all the elemnt of a list
- * @list_t: the list to be printed
- * 
+ * @h: The head of the list
+ *
  * Return: Number of Element printed
 */
 
-int print_list(const list_t *h)
+size_t print_list(const list_t *h)
 {
-    int n = 0;
-    list_t *current;
+int n = 0;
+list_t *current;
 
-    current = (list_t *)malloc(sizeof(list_t));
-    *current = *h;
-
-    while (current != NULL)
-    {
-        if (current->str == NULL)
-        {
-            printf("[0] (nil)\n");
-        }
-        else
-        {
-            printf("[%d] %s\n", current->len, current->str);
-        }
-        current = current->next;
-        n++;
-    }
-    free(current);
-    
-    return (n);
+current = (list_t *)malloc(sizeof(list_t));
+*current = *h;
+while (current != NULL)
+{
+if (current->str == NULL)
+{
+printf("[0] (nil)\n");
+}
+else
+{
+printf("[%d] %s\n", current->len, current->str);
+}
+current = current->next;
+n++;
+}
+free(current);
+return (n);
 }
