+++ 0x12-singly_linked_lists/lists.h
@@ -18,7 +18,7 @@ typedef struct list_s
     struct list_s *next;
 } list_t;
 
-int print_list(const list_t *h);
+size_t print_list(const list_t *h);
 int _putchar(char c);
 
 #endif
